refactor(ui): extract BrandSettingsTab component from AccountTab
